requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.1
Markdown==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
